# Conflicts:
#	applications/plugins/Compliant/contact/BaseContact.h
#	applications/plugins/Compliant/contact/CompliantContact.h
#	applications/plugins/Compliant/contact/FrictionCompliantContact.h
#	applications/plugins/Compliant/contact/PenaltyCompliantContact.h
#	applications/plugins/Compliant/mapping/ContactMapping.h
#	applications/plugins/Compliant/mapping/DifferenceMapping.h
#	applications/plugins/Compliant/odesolver/CompliantImplicitSolver.cpp
#	applications/plugins/SofaPython/python/SofaPython/noPicking.py
#	applications/plugins/image/ImageViewer.h
#	applications/sofa/gui/qt/viewer/qt/QtViewer.cpp
#	framework/sofa/core/Mapping.inl
#	modules/SofaBaseMechanics/UniformMass.cpp
#	modules/SofaBoundaryCondition/FixedConstraint.h
#	modules/SofaBoundaryCondition/FixedConstraint.inl
#	modules/SofaMeshCollision/MeshNewProximityIntersection.inl
#	modules/SofaMeshCollision/TriangleModel.inl
#	modules/SofaOpenglVisual/OglTexturePointer.cpp
#	modules/SofaOpenglVisual/OglTexturePointer.h
#	modules/sofa/simulation/common/MechanicalVisitor.cpp
#	modules/sofa/simulation/common/MechanicalVisitor.h


Former-commit-id: f3446a82a3367668e7ce0777ba07cdc68be1c4cf

+++ modules/SofaBaseCollision/SphereModel.inl
@@ -1,6 +1,6 @@
 /******************************************************************************
 *       SOFA, Simulation Open-Framework Architecture, development version     *
-*                (c) 2006-2015 INRIA, USTL, UJF, CNRS, MGH                    *
+*                (c) 2006-2016 INRIA, USTL, UJF, CNRS, MGH                    *
 *                                                                             *
 * This library is free software; you can redistribute it and/or modify it     *
 * under the terms of the GNU Lesser General Public License as published by    *
@@ -35,7 +35,6 @@
 #include <sofa/core/visual/VisualParams.h>
 #include <SofaBaseCollision/CubeModel.h>
 #include <sofa/core/ObjectFactory.h>
-#include <sofa/helper/system/gl.h>
 
 #include <sofa/core/topology/BaseMeshTopology.h>
 #include <sofa/simulation/common/Simulation.h>
@@ -113,7 +112,7 @@ void TSphereModel<DataTypes>::draw(const core::visual::VisualParams* vparams,int
 {
     TSphere<DataTypes> t(this,index);
 
-	vparams->drawTool()->drawSphere(t.p(), t.r());
+    vparams->drawTool()->drawSphere(t.p(), (float)t.r());
 }
